Extract errors to own file
